FOLSPRINGB-95 bugfix
