+++ src/test/java/org/folio/des/service/JobUpdatesServiceTest.java
@@ -20,6 +20,7 @@ import org.folio.des.domain.dto.JobStatus;
 import org.folio.des.domain.dto.Progress;
 import org.folio.des.domain.dto.VendorEdiOrdersExportConfig;
 import org.folio.des.repository.JobDataExportRepository;
+import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.integration.XOkapiHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -32,14 +33,30 @@ import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.ExitStatus;
 
 @ExtendWith(MockitoExtension.class)
-class JobUpdatesServiceTest  {
+class JobUpdatesServiceTest {
 
   @Mock
   private JobDataExportRepository repository;
   @InjectMocks
   private JobUpdatesService updatesService;
+  private JobUpdatesListenerService updatesListenerService;
 
   protected Map<String, Object> okapiHeaders;
+
+  @BeforeEach
+  void init() {
+    updatesListenerService = new JobUpdatesListenerService(new FolioModuleMetadata() {
+      @Override
+      public String getModuleName() {
+        return "mod-data-export-spring";
+      }
+
+      @Override
+      public String getDBSchemaName(String tenantId) {
+        return tenantId + "_mod_data_export_spring";
+      }
+    }, updatesService);
+  }
 
   @BeforeEach
   void setUp() {
@@ -52,7 +69,7 @@ class JobUpdatesServiceTest  {
 
   @Test
   @DisplayName("Update job with change")
-    void updateJobWithChange() {
+  void updateJobWithChange() {
     VendorEdiOrdersExportConfig config = new VendorEdiOrdersExportConfig();
     ExportTypeSpecificParameters parameters = new ExportTypeSpecificParameters();
 
@@ -85,7 +102,7 @@ class JobUpdatesServiceTest  {
     doReturn(Optional.of(job)).when(repository).findById(id);
     doReturn(job).when(repository).save(any());
 
-    updatesService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
+    updatesListenerService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
 
     verify(repository, times(1)).save(any());
   }
@@ -118,7 +135,7 @@ class JobUpdatesServiceTest  {
     doReturn(Optional.of(job)).when(repository).findById(id);
     doReturn(job).when(repository).save(any());
 
-    updatesService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
+    updatesListenerService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
 
     verify(repository, times(1)).save(any());
   }
@@ -159,7 +176,7 @@ class JobUpdatesServiceTest  {
 
     doReturn(Optional.of(job)).when(repository).findById(id);
 
-    updatesService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
+    updatesListenerService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
 
     verify(repository).save(isA(Job.class));
   }
@@ -173,7 +190,7 @@ class JobUpdatesServiceTest  {
 
     doReturn(Optional.empty()).when(repository).findById(id);
 
-    updatesService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
+    updatesListenerService.receiveJobExecutionUpdate(updatedJob, okapiHeaders);
 
     verify(repository, never()).save(any());
   }
